Fix GCC warning about stringop-overflow
When building with GCC, a warning is raised about string operation
overflow for the usages of `strncat`. This doesn't happen by default,
but when building on Arch Linux system, the `makepkg` default
configuration enables certain extra warnings.

After some searching it seems that GCC warns about it because calling
`strncat` with the source string length as 3rd argument is essentially
the same as calling the simpler `strcat` function.